reorganizing into more managable pieces
