+++ examples/kefir.js
@@ -1,0 +1,3 @@
+var streamSift = require('../lib')
+
+//kefir examples
