+++ examples/util.js
@@ -1,0 +1,1 @@
+var Kefir = require('Kefir')
